moves v2 services to legacy->two folder
